requirements.txt
discord.py==1.7.3
revChatGPT==0.3.1
